requirements.txt
torch==2.4.0
torchvision==0.19.0
decord==0.6.0
diffusers==0.30.1
numpy==1.26.0
gradio>=5.5.0
imageio>=2.35.1
imageio-ffmpeg>=0.5.1
Pillow==9.5.0
transformers==4.44.2
einops==0.8.0
opencv_python
